+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ### Fixed
 - Fix creator id being read from the wrong argument in JS runtime group update function.
 - Fix max count being incorrectly validated in group update JS runtime function.
+- Fix error handling when attempting to write records to a tournament that does not exist.
 
 ## [3.4.0] - 2021-07-08
 ### Added